jake - A lot of error validation

+++ src/java/com/myapp/struts/RemoveFromCartAction.java
@@ -12,6 +12,7 @@ import org.apache.struts.action.ActionErrors;
 import org.apache.struts.action.ActionForm;
 import org.apache.struts.action.ActionForward;
 import org.apache.struts.action.ActionMapping;
+import org.apache.struts.action.ActionMessage;
 
 /**
  *
@@ -20,7 +21,7 @@ import org.apache.struts.action.ActionMapping;
 public class RemoveFromCartAction extends org.apache.struts.action.Action {
 
     /* forward name="success" path="" */
-    private static final String SUCCESS = "success";
+    private static final String SUCCESS = "success",FAILURE="failure";
 
     /**
      * This is the action called from the Struts framework.
@@ -40,20 +41,38 @@ public class RemoveFromCartAction extends org.apache.struts.action.Action {
         
         FilmDAO dao = new FilmDAO();
         FilmForm info = (FilmForm) form;
-        
-        
+        boolean emptyBox,inCartCheck;
+        int error = 0;
         String f, u;
         
         HttpSession ses = request.getSession();
         ses.setAttribute("title", info.getTitle());       
         f = (String) ses.getAttribute("title");      
         
-        u = (String) ses.getAttribute("sessID");    
+        u = (String) ses.getAttribute("sessID");
+        
+        emptyBox = dao.textBoxEmptyCheck(f);
+        inCartCheck = dao.isFilmInCart(f, u);
+        if(emptyBox){
+            error = 4;
+        }
+        if(inCartCheck){
+            error = 3;
+        }
+        
+        if(!emptyBox && !inCartCheck){
         dao.removeFromCart(f, u);
-       
-        
-        
-        
         return mapping.findForward(SUCCESS);
+        }
+        else{
+            if(error == 4){
+            errors.add("title", new ActionMessage("errors.emptyBox"));
+            }
+            if(error == 3){
+               errors.add("title", new ActionMessage("errors.filmNotInCart")); 
+            }
+            this.saveErrors(request, errors);
+            return mapping.findForward(FAILURE);
+        }
     }
 }
